requirements.txt
PyQt6>=6.0.0
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.7.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
pydantic>=1.8.0,<1.9.0